Rename drop to shift
This commit renames "drop" to "shift", to bring it more in line with
"weak"/"lift".

The name "drop" was used to resolve the conflict with the _⊆_
constructor "weak", and stems from the "Normalization for Fitch-Style
Modal Calculi" mechanization which used the (inverted) drop/keep
terminology instead of weak/lift. As such, "weakRpl" would have also
been an appropriate name.

Closes #5